Create Fractional_Knapsack.cpp
Aim:
In Fractional Knapsack, we can break items for maximizing the total value of knapsack. This problem in which we can break an item is also called the fractional knapsack problem.

Analysis:
FractionalKnapsack has time complexity O(NlogN) where N is the number of items in S. The book suggests assuming S is a heap-based priority queue and then the removal has complexity Θ(logN) so the up to N removals take O(NlogN). The rest of the algorithm is O(N).